requirements.txt
boto3
python-dotenv
pydantic-ai[bedrock]